Migrate images to Imgix
Using imgix service to serve images replaces the existing cloudfront CDN for images
and replaces it with the Imgix cdn as well as their backend which allows for images
to be manipulated on the fly. This allows us to take a S3 bucket of 'master' image
and apply a watermark and resize them for the given device as required.

This should improve the overall performance of the system.